Remove CMB2 as a requirement
In order to avoid potential package duplication and/or version conflicts, the
burden of requiring the canonical CMB2 package should be left solely with the
project composer.json and not with any CMB2 extension manifest. For example, if
a project is including CMB2 as a must-use plugin, having other third-party
packages require it can result in the package being installed twice: as both
a regular and a must-use plugin.